requirements.txt
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.9.0
beautifulsoup4==4.13.4
bs4==0.0.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
databases==0.9.0
dnspython==2.7.0
exceptiongroup==1.2.2
fastapi==0.115.12
greenlet==3.2.3
gTTS==2.5.4
h11==0.14.0
idna==3.10
iniconfig==2.1.0
joblib==1.4.2
logzero==1.7.0
motor==3.7.0
narwhals==1.32.0
numpy==1.25.2
packaging==24.2
pandas==2.0.3
plotly==6.0.0
pluggy==1.6.0
pydantic==2.11.1
pydantic_core==2.33.0
Pygments==2.19.1
pymongo==4.11.1
pyotp==2.9.0
pytest==8.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.1
PyYAML==6.0.2
requests==2.31.0
scikit-learn==1.6.1
scipy==1.15.2
six==1.17.0
smartapi-python==1.5.4
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
starlette==0.46.1
tabulate==0.9.0
threadpoolctl==3.6.0
tomli==2.2.1
typing-inspection==0.4.0
typing_extensions==4.13.0
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
websocket-client==1.6.4
websockets==15.0.1
